Refactor Problem and Mapping modules
Problem:
- Split problems.jl into horizon/problem_interface/control_problems/
  abstraction_problems/recipes files.
- Add ProblemType -> ControlProblem/AbstractionProblem hierarchy; trajectory_success
  is now control-only (drop the =false fallback that masked "not applicable").
- Collapse discretize_problem to one generic ControlProblem method + horizon_round_up
  hook + reflective remake; loosen to ::Real.
- Rename abstraction-problem field region -> state_set; Infinity gains isinf.

Mapping:
- Split abstract_state_set.jl into state_sets/ folder; document MappedStateSet as the
  public surface.
- Dedup get_states_from_set to one generic delegating to _strict; uniform error stubs.
- Rename is_state_cover -> has_overlapping_cells, MappingSet -> FullStateSet.
- Disambiguate mapping-level add_set! -> cover! from the state-set add_set!.
- Fix allocating ellipsoid get_all_pos_by_coord (tuple product vs repeat).
- Give the hierarchical mapping a proper AbstractMultiLevelMapping supertype (was
  falsely <: GridMapping with a leaky (level,pos) position); rename file + add test.

Gate: problem 40/40; all mapping suites incl. new multi_level_mapping 12/12; symbolic;
UGA units + cosafe-LTL; UniformEllipsoid 21/21; Hybrid; regression fingerprint 9705631
unchanged; Aqua 11/11; format clean.
