+++ test/mapping/periodic_grid_mapping.jl
@@ -123,7 +123,7 @@ end
     pm = MP.PeriodicGridMapping(periodic_dims, periods, start, base)
 
     rect = UT.box(SVector(0.0, 0.0), SVector(2.0, 0.0))
-    MP.add_set!(pm, rect, MP.OUTER)
+    MP.cover!(pm, rect, MP.OUTER)
 
     # should have inserted positions (0,0), (1,0), (2,0) (mod 4 in dim 1)
     @test MP.is_valid_pos(pm, (0, 0))
