+++ test/problem/problems.jl
@@ -32,17 +32,18 @@ MathematicalSystems.stateset(s::DummySystem) = s.X
     @testset "AlternatingSimulationProblem fields" begin
         X = UT.box(SVector(-1.0, -1.0), SVector(1.0, 1.0))
         sys = DummySystem(X)
-        region = UT.box(SVector(-0.5, -0.5), SVector(0.5, 0.5))
+        state_set = UT.box(SVector(-0.5, -0.5), SVector(0.5, 0.5))
 
-        p = PR.AlternatingSimulationProblem(sys, region)
+        p = PR.AlternatingSimulationProblem(sys, state_set)
 
         @test p isa PR.ProblemType
+        @test p isa PR.AbstractionProblem
         @test p.system === sys
-        @test p.region == region
+        @test p.state_set == state_set
 
         # problems are immutable value objects
-        region2 = UT.box(SVector(-0.2, -0.2), SVector(0.2, 0.2))
-        @test_throws ErrorException p.region = region2
+        state_set2 = UT.box(SVector(-0.2, -0.2), SVector(0.2, 0.2))
+        @test_throws ErrorException p.state_set = state_set2
     end
 
     @testset "OptimalControlProblem fields" begin
@@ -110,6 +111,31 @@ MathematicalSystems.stateset(s::DummySystem) = s.X
         @test p.spec == spec
         @test p.labeling == lab
         @test p.ap_semantics == ap_sem
+    end
+
+    @testset "Problem hierarchy and interface" begin
+        X = UT.box(SVector(-2.0, -2.0), SVector(2.0, 2.0))
+        sys = DummySystem(X)
+        XI = UT.box(SVector(-1.0, -1.0), SVector(-0.5, -0.5))
+        XT = UT.box(SVector(0.5, 0.5), SVector(1.0, 1.0))
+
+        p = PR.OptimalControlProblem(sys, XI, XT, x -> 0.0, (x, u) -> 1.0, 5)
+        @test p isa PR.ControlProblem
+        @test !(p isa PR.AbstractionProblem)
+
+        # reach-avoid rounds the horizon down; the others round up
+        @test PR.horizon_round_up(p) == false
+        XS = UT.box(SVector(-1.5, -1.5), SVector(1.5, 1.5))
+        @test PR.horizon_round_up(PR.SafetyProblem(sys, XI, XS, 5)) == true
+
+        # remake swaps fields, re-infers type parameters, copies the rest
+        XT2 = UT.box(SVector(0.6, 0.6), SVector(1.1, 1.1))
+        p2 = PR.remake(p; target_set = XT2, time = 9)
+        @test p2 isa PR.OptimalControlProblem
+        @test p2.target_set == XT2
+        @test p2.time == 9
+        @test p2.initial_set === p.initial_set
+        @test p2.system === p.system
     end
 
     @testset "Plots/recipes smoke tests" begin
